requirements.txt
aiohttp==3.12.15
beautifulsoup4==4.13.4
lxml==6.0.0
playwright==1.54.0
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
schedule==1.2.2
